build(back): Gradle 빌드 로직 분리
